Updated SwitchStation2 to use only the incoming port # and infer the IP from the connection Updated SkullTalkService to only send incoming port # Updated SwitchStationClient to only send incoming port
Known bugs: Invalid Request Type is frequently made. Can't figure out why they're not lining up.
Also, the server is a different IP iff it's running on my phone; this is so that we can easily network the two (127.0.0.1 will not work with the phone, it is independent on the network). Currently updated to my router's phone number.

git-svn-id: http://skullduggery.googlecode.com/svn/trunk@114 edacab16-c986-11de-9e96-3193d784f102